requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
sqlparse==0.4.4
threadpoolctl==3.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
whitenoise==6.6.0
